requirements.txt
requests
python-dotenv
loguru